Add dart streams and improve futures
This commit introduces a new streams.dart file with different stream methods and updates futures.dart file as well. More specifically, the included stream methods generate random integers and return zero values in certain conditions. The improvements in the futures.dart entail addition of new futures that wait for specified delays and an error-prone calculation method that uses a completer to handle success and error.