Add macOS-specific configuration and application state classes
This commit introduces new files mainly related to macOS-specific configuration for the application. It also includes the creation of `MyAppState` and `MyAppStateWidget` classes which handle the app's state. These implementations are essential for maintaining and updating state throughout the application. Other notable additions are the test widget configurations and some flutter extension classes.